backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
watchdog==3.0.0
web3==6.11.0
requests==2.31.0
py-solc-x==2.0.4
python-dotenv==1.0.0
